+++ lapce-ui/src/lib.rs
@@ -5,6 +5,7 @@ pub mod alert;
 pub mod app;
 pub mod button;
 pub mod completion;
+pub mod dropdown;
 pub mod editor;
 pub mod explorer;
 pub mod find;