+++ lapce-ui/src/settings.rs
@@ -23,9 +23,12 @@ use lapce_data::{
         CommandExecuted, CommandKind, LapceUICommand, LAPCE_COMMAND,
         LAPCE_UI_COMMAND,
     },
-    config::{CoreConfig, EditorConfig, LapceTheme, TerminalConfig, UIConfig},
+    config::{
+        CoreConfig, DropdownInfo, EditorConfig, LapceTheme, TerminalConfig, UIConfig,
+    },
     data::{FocusArea, LapceEditorData, LapceTabData},
     document::{BufferContent, Document},
+    dropdown::DropdownData,
     keypress::KeyPressFocus,
     settings::{LapceSettingsFocusData, LapceSettingsKind, SettingsValueKind},
 };
@@ -34,6 +37,7 @@ use serde::Serialize;
 
 use crate::{
     button::Button,
+    dropdown::DropdownSelector,
     editor::view::LapceEditorView,
     keymap::LapceKeymap,
     scroll::{LapcePadding, LapceScroll},
@@ -439,6 +443,8 @@ impl LapceSettings {
                                     value = v.clone();
                                 }
                             }
+                            // TODO: let plugins define dropdowns
+                            let value = SettingsValue::from(value);
                             self.children.push(create_settings_item(
                                 data,
                                 volt.name.clone(),
@@ -457,7 +463,14 @@ impl LapceSettings {
         for (field, desc) in fields.iter().zip(descs.iter()) {
             // TODO(dbuga): we should generate kebab-case field names
             let field = field.replace('_', "-");
-            let value = settings.remove(&field).unwrap();
+            let value = if let Some(dropdown) =
+                data.config.get_dropdown_info(kind, &field)
+            {
+                SettingsValue::Dropdown(dropdown)
+            } else {
+                let value = settings.remove(&field).unwrap();
+                SettingsValue::from(value)
+            };
             self.children.push(create_settings_item(
                 data,
                 kind.to_string(),
@@ -552,6 +565,32 @@ struct LapceSettingsItemKeypress {
     cursor: usize,
 }
 
+#[derive(Debug, Clone)]
+pub enum SettingsValue {
+    Float(f64),
+    Integer(i64),
+    String(String),
+    Bool(bool),
+    Dropdown(DropdownInfo),
+    Empty,
+}
+impl From<serde_json::Value> for SettingsValue {
+    fn from(v: serde_json::Value) -> Self {
+        match v {
+            serde_json::Value::Number(n) => {
+                if n.is_f64() {
+                    SettingsValue::Float(n.as_f64().unwrap())
+                } else {
+                    SettingsValue::Integer(n.as_i64().unwrap())
+                }
+            }
+            serde_json::Value::String(s) => SettingsValue::String(s),
+            serde_json::Value::Bool(b) => SettingsValue::Bool(b),
+            _ => SettingsValue::Empty,
+        }
+    }
+}
+
 /// Create a settings item widget  
 /// Includes padding.
 fn create_settings_item(
@@ -559,34 +598,43 @@ fn create_settings_item(
     kind: String,
     key: String,
     desc: String,
-    value: serde_json::Value,
+    value: SettingsValue,
     event_sink: ExtEventSink,
 ) -> WidgetPod<LapceTabData, Box<dyn Widget<LapceTabData>>> {
     let insets = (10.0, 10.0);
     match value {
-        serde_json::Value::Number(n) => {
-            let value_kind = if n.is_f64() {
-                SettingsValueKind::Float
-            } else {
-                SettingsValueKind::Integer
-            };
-            WidgetPod::new(
-                LapcePadding::new(
-                    insets,
-                    InputSettingsItem::new(
-                        data,
-                        kind,
-                        key,
-                        desc,
-                        event_sink,
-                        n.to_string(),
-                        value_kind,
-                    ),
-                )
-                .boxed(),
+        // TODO: deduplicate these matches, the first three are just inputs with a string!
+        SettingsValue::Float(n) => WidgetPod::new(
+            LapcePadding::new(
+                insets,
+                InputSettingsItem::new(
+                    data,
+                    kind,
+                    key,
+                    desc,
+                    event_sink,
+                    n.to_string(),
+                    SettingsValueKind::Float,
+                ),
             )
-        }
-        serde_json::Value::String(s) => WidgetPod::new(
+            .boxed(),
+        ),
+        SettingsValue::Integer(n) => WidgetPod::new(
+            LapcePadding::new(
+                insets,
+                InputSettingsItem::new(
+                    data,
+                    kind,
+                    key,
+                    desc,
+                    event_sink,
+                    n.to_string(),
+                    SettingsValueKind::Integer,
+                ),
+            )
+            .boxed(),
+        ),
+        SettingsValue::String(s) => WidgetPod::new(
             LapcePadding::new(
                 insets,
                 InputSettingsItem::new(
@@ -601,16 +649,21 @@ fn create_settings_item(
             )
             .boxed(),
         ),
-        serde_json::Value::Bool(checked) => WidgetPod::new(
+        SettingsValue::Bool(checked) => WidgetPod::new(
             LapcePadding::new(
                 insets,
                 CheckBoxSettingsItem::new(key, kind, desc, checked),
             )
             .boxed(),
         ),
-        serde_json::Value::Array(_)
-        | serde_json::Value::Object(_)
-        | serde_json::Value::Null => WidgetPod::new(
+        SettingsValue::Dropdown(dropdown) => WidgetPod::new(
+            LapcePadding::new(
+                insets,
+                DropdownSettingsItem::new(data, key, kind, desc, dropdown),
+            )
+            .boxed(),
+        ),
+        SettingsValue::Empty => WidgetPod::new(
             LapcePadding::new(insets, EmptySettingsItem::new(key, kind, desc))
                 .boxed(),
         ),
@@ -1003,6 +1056,149 @@ impl Widget<LapceTabData> for CheckBoxSettingsItem {
                 2.0,
             );
         }
+    }
+}
+
+struct DropdownSettingsItem {
+    widget_id: WidgetId,
+    dropdown: WidgetPod<DropdownData<String, ()>, DropdownSelector<String, ()>>,
+
+    // We could have map of the key to a function to get the current items and selected item?
+    info: SettingsItemInfo,
+}
+impl DropdownSettingsItem {
+    fn new(
+        data: &mut LapceTabData,
+        key: String,
+        kind: String,
+        desc: String,
+        info: DropdownInfo,
+    ) -> Self {
+        let widget_id = WidgetId::next();
+
+        let dropdown = WidgetPod::new(DropdownSelector::default());
+
+        // Create our data and insert it into the settings data structure so we can access it later
+        let mut dropdown_data =
+            DropdownData::new(data.config.clone(), widget_id, ());
+        dropdown_data.update_from_info(info);
+
+        let settings = Arc::make_mut(&mut data.settings);
+        settings
+            .dropdown_data
+            .entry(kind.clone())
+            .or_insert_with(im::HashMap::new)
+            .insert(key.clone(), dropdown_data);
+
+        Self {
+            widget_id,
+            dropdown,
+
+            info: SettingsItemInfo::new(key, kind, desc),
+        }
+    }
+
+    /// Get the `DropdownData` for this widget, cloning it with the curent config
+    fn get_data_clone(&self, data: &LapceTabData) -> DropdownData<String, ()> {
+        data.settings
+            .dropdown_data
+            .get(&self.info.kind)
+            .unwrap()
+            .get(&self.info.key)
+            .unwrap()
+            .clone_with(data.config.clone())
+    }
+}
+impl Widget<LapceTabData> for DropdownSettingsItem {
+    fn id(&self) -> Option<WidgetId> {
+        Some(self.widget_id)
+    }
+
+    fn event(
+        &mut self,
+        ctx: &mut EventCtx,
+        event: &Event,
+        data: &mut LapceTabData,
+        env: &Env,
+    ) {
+        match event {
+            Event::Command(cmd) if cmd.is(LAPCE_UI_COMMAND) => {
+                let command = cmd.get_unchecked(LAPCE_UI_COMMAND);
+                if let LapceUICommand::DropdownItemSelected = command {
+                    // We just update the settings whenever you select an item in the dropdown
+                    // since it is unlikely that you're changing it fast enough to to be an
+                    // issue
+                    let dropdown_data = self.get_data_clone(data);
+                    if let Some(value) = dropdown_data.get_active_item() {
+                        self.info.update_settings(
+                            data,
+                            ctx,
+                            serde_json::json!(value),
+                        );
+                    }
+                }
+            }
+            _ => {}
+        }
+
+        let settings = Arc::make_mut(&mut data.settings);
+        let dropdown_data = settings
+            .dropdown_data
+            .get_mut(&self.info.kind)
+            .unwrap()
+            .get_mut(&self.info.key)
+            .unwrap();
+        dropdown_data.update_data(data.config.clone());
+
+        self.dropdown.event(ctx, event, dropdown_data, env);
+    }
+
+    fn lifecycle(
+        &mut self,
+        ctx: &mut LifeCycleCtx,
+        event: &LifeCycle,
+        data: &LapceTabData,
+        env: &Env,
+    ) {
+        let dropdown = self.get_data_clone(data);
+        self.dropdown.lifecycle(ctx, event, &dropdown, env);
+    }
+
+    fn update(
+        &mut self,
+        ctx: &mut UpdateCtx,
+        old_data: &LapceTabData,
+        data: &LapceTabData,
+        env: &Env,
+    ) {
+        let dropdown = self.get_data_clone(data);
+        self.dropdown.update(ctx, &dropdown, env);
+        self.info.update(ctx, old_data, data, env)
+    }
+
+    fn layout(
+        &mut self,
+        ctx: &mut LayoutCtx,
+        bc: &BoxConstraints,
+        data: &LapceTabData,
+        env: &Env,
+    ) -> Size {
+        let size = self.info.layout(ctx, bc, data, env, 0.0);
+
+        let dropdown = self.get_data_clone(data);
+        let dropdown_size = self.dropdown.layout(ctx, bc, &dropdown, env);
+        self.dropdown
+            .set_origin(ctx, &dropdown, env, Point::new(0.0, size.height));
+
+        ctx.set_paint_insets(4000.0);
+
+        Size::new(size.width, (size.height + dropdown_size.height).ceil())
+    }
+
+    fn paint(&mut self, ctx: &mut PaintCtx, data: &LapceTabData, env: &Env) {
+        self.info.paint(ctx, data, env, 0.0);
+        let dropdown = self.get_data_clone(data);
+        self.dropdown.paint(ctx, &dropdown, env);
     }
 }